fix: Merge with v5 and move codemods and tests
