--- modules/codemod/lib/space/spec/tsconfig.json
@@ -1,3 +1,0 @@
-{
-  "extends": "../../../../../tsconfig.spec.json"
-}
