+++ modules/codemod/lib/v5/index.ts
@@ -1,14 +1,25 @@
 import slashImports from './slashImports';
 import removeDefaultImports from './removeDefaultImports';
+// button codemods
 import removeButtonEnums from './removeButtonEnums';
 import renameButtonRefs from './renameButtonRefs';
+// spacing codemods
+import renameSpacingImports from './renameSpacingImports';
+import renameSpacingExpressionsAndTypeReferences from './renameSpacingExpressionsAndTypeReferences';
 
 import {API, FileInfo, Options} from 'jscodeshift';
 
 export default function transformer(file: FileInfo, api: API, options: Options) {
   // These will run in order. If your transform depends on others, place yours after dependent
   // transforms
-  const fixes = [slashImports, removeDefaultImports, removeButtonEnums, renameButtonRefs];
+  const fixes = [
+    slashImports,
+    removeDefaultImports,
+    removeButtonEnums,
+    renameButtonRefs,
+    renameSpacingImports,
+    renameSpacingExpressionsAndTypeReferences,
+  ];
 
   return fixes.reduce((source, fix) => fix({...file, source}, api, options), file.source);
 }
